# perf: (BREAKING) pre-allocate Writer buffer