--- a/src/index.ts
+++ b/src/index.ts
@@ -1,5 +1,5 @@
 import Reader from './lib/reader'
-import Writer from './lib/writer'
+import Writer, { WriterInterface } from './lib/writer'
 import Predictor from './lib/predictor'
 
-export { Reader, Writer, Predictor }
+export { Reader, Writer, WriterInterface, Predictor }