Optimize packet serialization/deserialization by reducing the number of `Buffer` allocations and avoiding bignum operations when numbers will suffice.

See also:
- https://github.com/interledgerjs/ilp-packet/pull/53
- https://github.com/interledgerjs/btp-packet/pull/34
- https://github.com/interledgerjs/ilp-protocol-stream/pull/97

## Breaking changes:

  - `Predictor` can no longer be used directly as a `Writer`. Instead, both `Predictor` and `Writer` implement `WriterInterface`.
  - `write`/`writeOctetString`/`writeVarOctetString` can no longer receive a `Writer`. Because the `Writer` doesn't keep a list of component buffers, there's no efficiency gain to this. `createVarOctetString` is faster, anyway.

## Features:

  - Add `Predictor.measureVarOctetString`, which is useful when calculating a `Writer`'s buffer size in advance.
  - Add `{Predictor,Writer}.createVarOctetString` (see below).

## Bug fixes:

  - `writeUInt`/`writeInt`: Fix length check for large bignums (>6 bytes).

## Optimizations:

### `Writer` / `Predictor`:

  - `writeVarUInt`/`writeVarInt`: avoid creating `BigNumber`s when possible (in both `Writer` and `Predictor`). Use `writeUInt`/`writeInt` to write numbers without allocating buffers or bignums.
  - Calculating the buffer size of numbers/bignums is much faster, and can usually avoid stringifying the number/bignumber (which is slow).
  - `createVarOctetString` is used to construct a pre-allocated OER-encoded VarOctetString (replacing `writeVarOctetString(otherWriter)`).
  - A `Writer` constructor can be passed an exact capacity, which will allow it to only allocate a buffer once. If no capacity is provided, it will allocate a small initial buffer, then resize when it is filled. The former method is fastest by far, but the latter is still faster than the old strategy of constructing a list of buffers and concatenating them at the end.

### `Reader`:

  - `readLengthPrefix`: `use `readUInt8Number()` instead of `readUInt8BigNum().toNumber()` to get the length.

## Benchmarks
(v0: before; v1: after):

### ilp-packet

```
serializeIlpPrepare      v0 x 224,294 ops/sec ±0.84% (91 runs sampled)
serializeIlpPrepare      v1 x 418,535 ops/sec ±0.88% (91 runs sampled)
deserializeIlpPrepare    v0 x 236,698 ops/sec ±1.28% (91 runs sampled)
deserializeIlpPrepare    v1 x 322,443 ops/sec ±0.45% (91 runs sampled)
serializeIlpFulfill      v0 x 537,387 ops/sec ±0.86% (90 runs sampled)
serializeIlpFulfill      v1 x 1,054,131 ops/sec ±1.20% (88 runs sampled)
deserializeIlpFulfill    v0 x 1,079,358 ops/sec ±0.68% (94 runs sampled)
deserializeIlpFulfill    v1 x 4,630,731 ops/sec ±1.60% (89 runs sampled)
serializeIlpReject       v0 x 260,279 ops/sec ±0.82% (88 runs sampled)
serializeIlpReject       v1 x 457,782 ops/sec ±0.98% (91 runs sampled)
deserializeIlpReject     v0 x 478,236 ops/sec ±0.69% (97 runs sampled)
deserializeIlpReject     v1 x 1,333,481 ops/sec ±1.00% (89 runs sampled)
```

### btp-packet:
(benchmarks a `Message` packet with a single `ilp` subprotocol)

```
serialize:   Message     v0 x 250,978 ops/sec ±0.85% (91 runs sampled)
serialize:   Message     v1 x 862,162 ops/sec ±1.08% (94 runs sampled)
deserialize: Message     v0 x 735,601 ops/sec ±1.88% (94 runs sampled)
deserialize: Message     v1 x 2,605,169 ops/sec ±0.60% (92 runs sampled)
```

### ilp-protocol-stream:
(note that these benchmarks don't include encryption/decryption)

```
serialize:   MoneyFrame          v0 x 25,560 ops/sec ±0.84% (93 runs sampled)
serialize:   MoneyFrame          v1 x 116,407 ops/sec ±1.60% (95 runs sampled)
deserialize: MoneyFrame          v0 x 114,379 ops/sec ±2.44% (89 runs sampled)
deserialize: MoneyFrame          v1 x 206,280 ops/sec ±0.55% (90 runs sampled)
```